backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.7.1
python-binance==1.0.19
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.20
